HelpCategoriesScreen done

+++ app/src/main/java/com/example/practicaltasksmvp/data/DataManager.kt
@@ -95,7 +95,7 @@ class DataManager(private val context: Context, private val apiManager: ApiManag
                 .map { mapToEntity(context, it) }
         } else {
             return data
-                .find { helpCategoryPojo -> helpCategoryPojo.id.equals(categoryId) }
+                .find { helpCategoryPojo -> helpCategoryPojo.id.toLong().equals(categoryId) }
                 ?.articles?.map { mapToEntity(context, it) }
         }
     }
@@ -104,7 +104,7 @@ class DataManager(private val context: Context, private val apiManager: ApiManag
         return data
             .flatMap { helpCategoryPojo -> helpCategoryPojo.articles }
             .map { mapToEntity(context, it) }
-            .find { newsArticleEntity -> newsArticleEntity.id.equals(articleId) }
+            .find { newsArticleEntity -> newsArticleEntity.id.toLong().equals(articleId) }
     }
 
     private fun extractCategories(data: List<HelpCategoryPojo>): List<HelpCategoryEntity> {
